requirements.txt
streamlit
openai
pdfminer.six
streamlit
openai
pymupdf
python-dotenv
numpy
